requirements.txt
keyboard==0.13.5
Pillow==9.2.0
windows_toasts==0.4.0
